requirements.txt
aiocache==0.11.1
aiohttp==3.6.2
aiohttp-remotes==0.1.2
aiomcache==0.6.0
aioredis==1.3.1
async-timeout==3.0.1
attrs==20.2.0
chardet==3.0.4
gunicorn==20.0.4
hiredis==1.1.0
idna==2.10
idna-ssl==1.1.0
msgpack==1.0.0
multidict==4.7.6
psutil==5.7.2
PyYAML==5.3.1
statsd==3.3.0
typing-extensions==3.7.4.3
yarl==1.5.1
